Rename localization diagnostics files to be consistent
