+++ src/diagnostics/LocalizationXML.ts
