+++ src/library/bg3/schema/localization.xml.ts
